+++ meta_package_manager/tests/test_cli.py
@@ -25,12 +25,11 @@ import click
 import pytest
 import simplejson as json
 from boltons.iterutils import flatten
-from boltons.strutils import strip_ansi
 from cli_helpers.tabular_output import TabularOutputFormatter
 
-from .. import __version__, logger
-from ..cli import RENDERING_MODES, WINDOWS_MODE_BLACKLIST
 from .conftest import MANAGER_IDS, run_cmd, unless_windows
+from .. import __version__, logger, strip_ansi
+from ..cli import RENDERING_MODES, WINDOWS_MODE_BLACKLIST, cli
 
 """ Common tests for all CLI basic features and templates for subcommands. """